requirements.txt
# MMSU Prior Art Search Tool Requirements
# Core Flask Dependencies
Flask==3.1.0
Flask-SQLAlchemy==3.1.2
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-WTF==1.2.2
WTForms==3.1.2

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25

# Production Server
gunicorn==21.2.0

# PDF Generation
WeasyPrint==62.3
Pillow==10.2.0

# File Processing
python-docx==1.1.2
PyPDF2==3.0.1
python-magic==0.4.27

# API Integration
requests==2.32.3

# Security & Validation
email-validator==2.2.0
python-dotenv==1.0.1

# Utilities
python-dateutil==2.8.2

# Development Dependencies
pytest==8.2.2
pytest-flask==1.3.0
pytest-cov==5.0.0
black==24.4.2
flake8==7.1.0

# Optional: Redis for session management (if needed)
# redis==5.0.8
# Flask-Session==0.8.0
